Moved Services Enum to Config to prevent circular import
