+++ modules/StampyControls.py
@@ -1,9 +1,8 @@
 import sys
 import discord
 from modules.module import Module, Response
-from config import stampy_control_channel_names, TEST_RESPONSE_PREFIX
+from config import stampy_control_channel_names, TEST_RESPONSE_PREFIX, Services
 from utilities import Utilities, get_github_info, get_memory_usage, get_running_user_info, get_question_id
-from utilities.serviceutils import Services
 
 
 class StampyControls(Module):
